refactor: StationGraph 인터페이스 수정
- 실제 클라이언트클래스에서 필요한 부분만 interface로 작성
- 구체적인 부분은 외부라이브러리를 구현하는 쪽에서 하도록 수정
