+++ src/main/java/nextstep/subway/path/domain/jgrapht/WeightedMultiStationGraph.java
@@ -1,6 +1,9 @@
 package nextstep.subway.path.domain.jgrapht;
 
+import java.util.List;
 import nextstep.subway.exception.NotFoundException;
+import nextstep.subway.line.domain.Line;
+import nextstep.subway.line.domain.Section;
 import nextstep.subway.path.domain.Path;
 import nextstep.subway.path.domain.StationGraph;
 import nextstep.subway.station.domain.Station;
@@ -8,7 +11,9 @@ import org.jgrapht.GraphPath;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.DefaultWeightedEdge;
 import org.jgrapht.graph.WeightedMultigraph;
+import org.springframework.stereotype.Component;
 
+@Component
 public class WeightedMultiStationGraph implements StationGraph {
     private WeightedMultigraph<Station, DefaultWeightedEdge> graph;
 
@@ -17,13 +22,12 @@ public class WeightedMultiStationGraph implements StationGraph {
     }
 
     @Override
-    public void addVertex(Station s) {
-        graph.addVertex(s);
-    }
-
-    @Override
-    public void addEdgeWithDistance(Station source, Station target, Integer distance) {
-        graph.setEdgeWeight(graph.addEdge(source, target), distance);
+    public StationGraph createGraph(List<Line> lines) {
+        for (Line line: lines) {
+            addVertex(line.getStations());
+            addEdgeWeight(line.getSections());
+        }
+        return this;
     }
 
     @Override
@@ -39,5 +43,17 @@ public class WeightedMultiStationGraph implements StationGraph {
             throw new NotFoundException("점접이 없습니다.");
         }
         return Path.of(path.getVertexList(), Double.valueOf(path.getWeight()).intValue());
+    }
+
+    private  void addVertex(List<Station> stations) {
+        for (Station s: stations) {
+            graph.addVertex(s);
+        }
+    }
+
+    private void addEdgeWeight(List<Section> sections) {
+        for (Section s: sections) {
+            graph.setEdgeWeight(graph.addEdge(s.getUpStation(), s.getDownStation()), s.getDistance().get());
+        }
     }
 }
